Renaming utils to camelCase
