+++ lib/InputTest.ts
@@ -14,13 +14,13 @@ import {
 } from './polyfills';
 import { getDefaultDevices } from './polyfills/enumerateDevices';
 import { TimeMeasurement } from './types';
-import { determineSilent } from './utils/DetermineSilent';
+import { detectSilence } from './utils';
 import {
   InvalidityRecord,
   validateDeviceId,
   validateOptions,
   validateTime,
-} from './utils/OptionValidation';
+} from './utils/optionValidation';
 
 export declare interface InputTest {
   /**
@@ -206,7 +206,7 @@ export class InputTest extends EventEmitter {
     this._cleanup();
 
     this._endTime = Date.now();
-    const didPass: boolean = pass && !determineSilent(this._values);
+    const didPass: boolean = pass && !detectSilence(this._values);
     const report: InputTest.Report = {
       deviceId: this._options.deviceId || (
         this._defaultDevices.audioinput &&
