+++ lib/utils/optionValidation.ts
